feat(ingest): extract smart-ingest service — foundation for unified upload (PR A)
Lifts the body of uploadAndAnalyze + finalizeDocument into a new
shared service `lib/compliance/smart-ingest.ts` exposing two pure
async entry points:

  • analyzeAndDraft({ companyId, userId, filePath, fileName })
  • finalizeIngestion({ companyId, userId, documentId, confirmed })

The server actions in app/data-room/actions-document-agent.ts are
now thin wrappers that handle auth and delegate. **Zero behaviour
change at the tracker** — same draft → analyze → confirm → finalize
flow, same DB writes (CompanyDocument metadata, embedding refresh,
fact persistence, ComplianceFiling upsert).

Why a separate file: PR B will call analyzeAndDraft from the Supabase
Cron worker for vault uploads (background ingestion), PR D will call
it from onboarding for cert OCR. Auth-agnostic by design — callers
verify access BEFORE invoking and pass userId explicitly. Makes the
service unit-testable without HTTP context.

Foundation only. The new logic (requirement auto-resolution,
missing-period auto-generation, async progress) lands in subsequent
PRs in this chain.
